web-platform/backend/requirements.txt
# Backend MacTester Web Platform
# Python 3.8+

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Data Processing
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dateutil>=2.8.2

# MT5 Integration
MetaTrader5>=5.0.5388
psutil>=5.9.0

# File Handling
pyyaml==6.0.1
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Utilities
aiofiles==23.2.1
watchdog==3.0.0

# Backup
schedule==1.2.0

# Telegram (opcional)
python-telegram-bot>=20.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

